Fixes #60
The problem was that sometimes the system requests a redraw of the OpenGL view, and we were
assuming these requests were made in lockstep with AlgorithmStepper events.  We were clearing
the lists of elements to be rendered in the OpenGL views after the view was drawn; and
sometimes nobody generated a new list of elements before the the OpenGL view was prompted to
draw again.