Issue #87: added retry with perturbation for Delaunay insertion.
The problem occurs often when the large initial mesh is used, and perturbation doesn't seem to help.
An alternative solution (not implemented) is to detect if the new vertex lies on an existing edge, and
if so, split that edge to form two additional triangles (instead of splitting a face to form three).