Issue #142: slight change to quiescent updating of button enabled states
I tried removing the quiescent delay and instead adding a counter to EdObjectArray
that increments with each modification, and using this to avoid recalculating things related
to the objects (specifically the button enable states, which includes the (relatively
expensive) 'unhidePossible()' call).  This does work,
but for instance when the user is editing a segment endpoint, these calculations all need to
be done anyways.  I decreased the quiescent delay to 1/10 sec.